requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
colorama==0.4.6
creategitignore==0.2.1
Django==4.1.7
django-appconf==1.0.5
django-imagekit==4.1.0
django-jazzmin==2.6.0
pilkit==2.0
Pillow==9.4.0
python-decouple==3.8
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
utils==1.0.1
